rename Omni*Imagery files to Omni*RasterOverlay
