+++ src/core/include/cesium/omniverse/OmniRasterOverlay.h
